user reset and change password

+++ application/controllers/ProductsController.php
@@ -2,16 +2,15 @@
 
 class ProductsController extends Controller{
 
-    private function gotologin(){
-        $this->_template = new Template($this->_controller,'gotologin');
-        $this->set('title','GoToLogin');
-        $this->set('controller',$this->_controller);
-        $this->_template->render();
+    private function gotologin($error = 'previous'){
+        $this->_session->start();
+        $this->_session->put('error',"Login to access $error page !");
+        Router::go(array('users','login'));
     }
 
     public function viewall(){
         if(!$this->_session->isAuth()){
-            $this->gotologin();
+            $this->gotologin("Products");
         }
         else{
             $this->set('title','Products');
@@ -36,35 +35,39 @@ class ProductsController extends Controller{
         }
     }
     public function delete(){
-        $id = $_POST['id'];
-        $file = $_POST['file'];
-        $this->set('title','');
-        $this->Product->delete($id,$file);
-        //header("Location:". ROOT . '/' . $this->_controller .'/viewall');
-        //exit();
-        $this->_template->render();
+        if(!$this->_session->isAuth()){
+            $this->gotologin();
+        }
+        else {
+            $id = $_POST['id'];
+            $file = $_POST['file'];
+            $this->set('title', '');
+            $this->Product->delete($id, $file);
+            Router::go(array('products','viewall'));
+        }
     }
     public function save(){
-        if(isset($_POST['type'])&&isset($_POST['name'])&&isset($_POST['price'])) {
-            $this->Product->setId($_POST['id']);
-            $this->Product->setName($_POST['name']);
-            $this->Product->setPrice($_POST['price']);
-            $this->Product->setType($_POST['type']);
-            if(isset($_FILES["file"]["name"]))  $this->Product->setFile(basename($_FILES["file"]["name"]));
-            else $this->Product->setFile('default');
-
-            if ($this->Product->getId()) {
-                if($this->Product->update()) $this->set('succes',true);
-                else $this->set('succes',false);
-            }
-                else {
-                    if($this->Product->add_new()) $this->set('succes',true);
-                    else $this->set('succes',false);
-                }
+        if(!$this->_session->isAuth()){
+            $this->gotologin();
         }
+        else {
+            if (isset($_POST['type']) && isset($_POST['name']) && isset($_POST['price'])) {
+                $this->Product->setId($_POST['id']);
+                $this->Product->setName($_POST['name']);
+                $this->Product->setPrice($_POST['price']);
+                $this->Product->setType($_POST['type']);
+                if (isset($_FILES["file"]["name"])) $this->Product->setFile(basename($_FILES["file"]["name"]));
+                else $this->Product->setFile('default');
 
-        $path = Router::buildPath(array($this->_controller,'viewall'));
-        header("Location:" . $path);
-        exit();
+                if ($this->Product->getId()) {
+                    if ($this->Product->update()) $this->set('succes', true);
+                    else $this->set('succes', false);
+                } else {
+                    if ($this->Product->add_new()) $this->set('succes', true);
+                    else $this->set('succes', false);
+                }
+            }
+            Router::go(array($this->_controller, 'viewall'));
+        }
     }
 }
